+++ .vscode/extensions.json
@@ -2,7 +2,6 @@
     "recommendations": [
         "sumneko.lua",
         "rc.lmaobox-lua-api-snippets",
-        "gccfeli.vscode-lua",
         "notnoobmaster.lua-beautifier"
     ]
 }
