+++ .vscode/settings.json
@@ -16,5 +16,9 @@
                 "autoShowOutputPanel": "error"
             }
         ]
-    }
+    },
+    "Lua.workspace.userThirdParty": [
+        "C:/Users/Terminatort8000/AppData/Local/Lmaobox-Annotations-master/library"
+    ],
+    "stylua.targetReleaseVersion": "latest"
 }
