+++ BundleAndDeploy.bat
@@ -1,5 +1,5 @@
 @echo off
 
 node bundle.js
-move /Y "Lmaobot.lua" "%localappdata%"
+move /Y "MedBot.lua" "%localappdata%"
 exit
