+++ MedBot/Utils/Globals.lua
